requirements.txt
# Core framework
kivy==2.3.0
kivymd==1.2.0

# Maps and location
kivy-garden.mapview==1.0.6
plyer==2.1.0

# HTTP requests for Google APIs
requests==2.31.0

# Text-to-speech
pyttsx3==2.90
google-cloud-texttospeech==2.16.3  # For realistic neural voices (optional)
pygame==2.5.2  # For audio playback

# Build tools for mobile
buildozer==1.5.0  # For Android
kivy-ios==1.3.1   # For iOS
